Made createCSV that dumps all data in the database into a .CSV file.
Requirements: write permissions and the file already created.
ToDo: add exception handler, format the data before putting it into the .csv file i.e. adding category differentiation in the .csv file, and figuring out a way to have php create the file, add write permissions prior to writing to the file b/c if the file doesn't exist && it doesn't have write permission, it won't be able to save the file. Maybe add an interactive file handler to allow the user to specify where to put the file.